snapshots: Move implementation to lvs::Lvol
The pool layer rework in e6f7517 introduced lvs and is the preferred
home for new functionality on pools/replicas so move snapshot creation
there.
