+++ mayastor/src/bdev/nexus/nexus_bdev_snapshot.rs
@@ -5,7 +5,7 @@ use rpc::mayastor::CreateSnapshotReply;
 use crate::{
     bdev::nexus::nexus_bdev::{Error, Nexus},
     core::BdevHandle,
-    replica::Replica,
+    lvs::Lvol,
 };
 
 impl Nexus {
@@ -14,7 +14,7 @@ impl Nexus {
         if let Ok(h) = BdevHandle::open_with_bdev(&self.bdev, true) {
             match h.create_snapshot().await {
                 Ok(t) => Ok(CreateSnapshotReply {
-                    name: Replica::format_snapshot_name(&self.bdev.name(), t),
+                    name: Lvol::format_snapshot_name(&self.bdev.name(), t),
                 }),
                 Err(_e) => Err(Error::FailedCreateSnapshot),
             }
