+++ mayastor/src/subsys/nvmf/admin_cmd.rs
@@ -1,11 +1,13 @@
 //! Handlers for custom NVMe Admin commands
 
+use std::convert::TryFrom;
+
 use spdk_sys::{spdk_bdev, spdk_bdev_desc, spdk_io_channel, spdk_nvmf_request};
 
 use crate::{
     bdev::nexus::nexus_io::nvme_admin_opc,
     core::{Bdev, Reactors},
-    replica::Replica,
+    lvs::Lvol,
 };
 
 /// NVMf custom command handler for opcode c0h
@@ -42,17 +44,18 @@ extern "C" fn nvmf_create_snapshot_hdlr(req: *mut spdk_nvmf_request) -> i32 {
     }
 
     let bd = Bdev::from(bdev);
-    if let Some(replica) = Replica::from_bdev(&bd) {
+    let base_name = bd.name();
+    if let Ok(lvol) = Lvol::try_from(bd) {
         let cmd = unsafe { &*spdk_sys::spdk_nvmf_request_get_cmd(req) };
         let snapshot_time = unsafe {
             cmd.__bindgen_anon_1.cdw10 as u64
                 | (cmd.__bindgen_anon_2.cdw11 as u64) << 32
         };
         let snapshot_name =
-            Replica::format_snapshot_name(&bd.name(), snapshot_time);
+            Lvol::format_snapshot_name(&base_name, snapshot_time);
         // Blobfs operations must be on md_thread
         Reactors::master().send_future(async move {
-            replica.create_snapshot(req, &snapshot_name).await;
+            lvol.create_snapshot(req, &snapshot_name).await;
         });
         1 // SPDK_NVMF_REQUEST_EXEC_STATUS_ASYNCHRONOUS
     } else {
